Make idempotent the creation of -orig files.

+++ install_clouderadirector.sh
@@ -25,7 +25,11 @@ export http_proxy
 export https_proxy
 wget -q https://archive.cloudera.com/director/redhat/${OSREL}/x86_64/director/cloudera-director.repo -O /etc/yum.repos.d/cloudera-director.repo
 yum -y -e1 -d1 install cloudera-director-server cloudera-director-client
-cp -p /etc/cloudera-director-server/application.properties /etc/cloudera-director-server/application.properties-orig
+if [ ! -f /etc/cloudera-director-server/application.properties-orig ]; then
+  cp -p /etc/cloudera-director-server/application.properties /etc/cloudera-director-server/application.properties-orig
+else
+  cp -p /etc/cloudera-director-server/application.properties /etc/cloudera-director-server/application.properties.`date '+%Y%m%d%H%M%S'`
+fi
 chgrp cloudera-director /etc/cloudera-director-server/application.properties
 chmod 0640 /etc/cloudera-director-server/application.properties
 sed -i -e '/lp.encryption.twoWayCipher:/a\
